backend/requirements.txt
fastapi
uvicorn
sentence-transformers
numpy
PyPDF2
python-docx
google-generativeai
